ai-service/requirements.txt
# FastAPI and server dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0

# ML and NLP dependencies
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.1
transformers==4.29.2
torch==2.0.1
sentencepiece==0.1.99

# Testing dependencies
pytest==7.3.1
httpx==0.24.1

# Utilities
python-multipart==0.0.6